@@ -10,8 +10,7 @@
     "azure-iot-common": "1.1.6",
     "azure-iot-device": "1.1.10",
     "debug": "^2.6.0",
-    "mqtt": "^1.14.1",
-    "es5-shim": "^4.5.9"
+    "mqtt": "^1.14.1"
   },
   "devDependencies": {
     "chai": "^3.5.0",
@@ -19,18 +18,21 @@
     "jshint": "^2.9.4",
     "mocha": "^3.2.0",
     "sinon": "^1.17.7",
-    "tslint": "^3.14.0",
-    "typescript": "^1.8.10"
+    "tslint": "^5.1.0",
+    "typescript": "2.2.2",
+    "@types/node": "^7.0.5",
+    "@types/mqtt": "0.0.34"
   },
   "scripts": {
-    "lint": "jshint --show-non-errors . && tslint -c ../../../tslint.json \"{lib,.}/*.ts\"",
+    "lint": "tslint --type-check --project . -c ../../../tslint.json",
+    "build": "tsc",
     "unittest-min": "istanbul cover --report none node_modules/mocha/bin/_mocha -- --reporter dot test/_*_test.js",
     "alltest-min": "istanbul cover --report none node_modules/mocha/bin/_mocha -- --reporter dot test/_*_test*.js",
     "unittest": "istanbul cover node_modules/mocha/bin/_mocha -- --reporter spec test/_*_test.js",
     "alltest": "istanbul cover node_modules/mocha/bin/_mocha -- --reporter spec test/_*_test*.js",
-    "ci": "npm -s run lint && npm -s run alltest-min && npm -s run check-cover",
-    "test": "npm -s run lint && npm -s run unittest",
-    "check-cover": "istanbul check-coverage --statements 92 --branches 84 --functions 80 --lines 92"
+    "ci": "npm -s run lint && npm -s run build && npm -s run alltest-min && npm -s run check-cover",
+    "test": "npm -s run lint && npm -s run build && npm -s run unittest",
+    "check-cover": "istanbul check-coverage --statements 89 --branches 76 --functions 76 --lines 76"
   },
   "engines": {
     "node": ">= 0.10"
