@@ -8,10 +8,13 @@ export interface Transport extends EventEmitter {
   updateSharedAccessSignature(sharedAccessSignature: string, done: (err?: Error, result?: results.SharedAccessSignatureUpdated) => void): void;
   getReceiver(func: (err?: Error, receiver?: Receiver) => void): void;
   sendEvent(message: Message, done: (err?: Error, result?: results.MessageEnqueued) => void): void;
-  sendEventBatch(messages: Message[], done: (err: Error, result?: results.MessageEnqueued) => void): void;
   complete(message: Message, done: (err?: Error, result?: results.MessageCompleted) => void): void;
   reject(message: Message, done: (err?: Error, results?: results.MessageRejected) => void): void;
   abandon(message: Message, done: (err?: Error, results?: results.MessageAbandoned) => void): void;
+}
+
+export interface BatchingTransport extends Transport {
+  sendEventBatch(messages: Message[], done: (err: Error, result?: results.MessageEnqueued) => void): void;
 }
 
 export interface StableConnectionTransport extends Transport {
