@@ -1,11 +1,18 @@
 // Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
 'use strict';
-
-var Base = require('./mqtt.js');
-var util = require('util');
-
+var __extends = (this && this.__extends) || (function () {
+    var extendStatics = Object.setPrototypeOf ||
+        ({ __proto__: [] } instanceof Array && function (d, b) { d.__proto__ = b; }) ||
+        function (d, b) { for (var p in b) if (b.hasOwnProperty(p)) d[p] = b[p]; };
+    return function (d, b) {
+        extendStatics(d, b);
+        function __() { this.constructor = d; }
+        d.prototype = b === null ? Object.create(b) : (__.prototype = b.prototype, new __());
+    };
+})();
+Object.defineProperty(exports, "__esModule", { value: true });
+var mqtt_1 = require("./mqtt");
 /**
  * @class        module:azure-iot-device-mqtt.MqttWs
  * @classdesc    Provides MQTT over WebSockets transport for the device [client]{@link module:azure-iot-device.Client} class.
@@ -15,12 +22,15 @@ var util = require('util');
 /*Codes_SRS_NODE_DEVICE_MQTT_12_001: [The constructor shall accept the transport configuration structure.]*/
 /*Codes_SRS_NODE_DEVICE_MQTT_12_002: [The constructor shall store the configuration structure in a member variable.]*/
 /*Codes_SRS_NODE_DEVICE_MQTT_12_003: [The constructor shall create an base transport object and store it in a member variable.]*/
-function MqttWs(config) {
-  Base.call(this, config);
-  /*Codes_SRS_NODE_DEVICE_MQTT_16_017: [The `MqttWs` constructor shall initialize the `uri` property of the `config` object to `wss://<host>:443/$iothub/websocket`.]*/
-  this._config.uri  = 'wss://' + config.host + ':443/$iothub/websocket';
-}
-
-util.inherits(MqttWs, Base);
-
-module.exports = MqttWs;
+var MqttWs = (function (_super) {
+    __extends(MqttWs, _super);
+    function MqttWs(config) {
+        var _this = _super.call(this, config) || this;
+        /*Codes_SRS_NODE_DEVICE_MQTT_16_017: [The `MqttWs` constructor shall initialize the `uri` property of the `config` object to `wss://<host>:443/$iothub/websocket`.]*/
+        _this._config.uri = 'wss://' + config.host + ':443/$iothub/websocket';
+        return _this;
+    }
+    return MqttWs;
+}(mqtt_1.Mqtt));
+exports.MqttWs = MqttWs;
+//# sourceMappingURL=mqtt_ws.js.map
