@@ -5,7 +5,7 @@
 
 var assert = require('chai').assert;
 
-var MqttTwinReceiver = require('../lib/mqtt-twin-receiver.js');
+var MqttTwinReceiver = require('../lib/mqtt_twin_receiver.js').MqttTwinReceiver;
 var MqttProvider = require('./_fake_mqtt.js');
 var sinon = require('sinon');
 
